Extract some join tests to separate module
